+++ src/NodeAssn.java
@@ -3,16 +3,16 @@
 
 public class NodeAssn extends Node {
 
-    private String id;
-    private NodeExpr expr;
+	private String id;
+	private NodeExpr expr;
 
-    public NodeAssn(String id, NodeExpr expr) {
-	this.id=id;
-	this.expr=expr;
-    }
+	public NodeAssn(String id, NodeExpr expr) {
+		this.id = id;
+		this.expr = expr;
+	}
 
-    public int eval(Environment env) throws EvalException {
-	return env.put(id,expr.eval(env));
-    }
+	public double eval(Environment env) throws EvalException {
+		return env.put(id, expr.eval(env));
+	}
 
 }
