+++ src/NodeBlock.java
@@ -7,15 +7,15 @@ public class NodeBlock extends Node {
 		this.stmt = stmt;
 		this.block = block;
 	}
-	
+
 	@Override
-	public int eval(Environment env) throws EvalException {
-		int r = this.stmt.eval(env);
-		
+	public double eval(Environment env) throws EvalException {
+		double r = this.stmt.eval(env);
+
 		if (block != null) {
 			r = this.block.eval(env);
 		}
-		
+
 		return r;
 	}
