+++ src/NodeFactExpr.java
@@ -3,14 +3,14 @@
 
 public class NodeFactExpr extends NodeFact {
 
-    private NodeExpr expr;
+	private NodeExpr expr;
 
-    public NodeFactExpr(NodeExpr expr) {
-	this.expr=expr;
-    }
+	public NodeFactExpr(NodeExpr expr) {
+		this.expr = expr;
+	}
 
-    public int eval(Environment env) throws EvalException {
-	return expr.eval(env);
-    }
+	public double eval(Environment env) throws EvalException {
+		return expr.eval(env);
+	}
 
 }
