+++ src/NodeFactId.java
@@ -3,15 +3,15 @@
 
 public class NodeFactId extends NodeFact {
 
-    private String id;
+	private String id;
 
-    public NodeFactId(int pos, String id) {
-	this.pos=pos;
-	this.id=id;
-    }
+	public NodeFactId(int pos, String id) {
+		this.pos = pos;
+		this.id = id;
+	}
 
-    public int eval(Environment env) throws EvalException {
-	return env.get(pos,id);
-    }
+	public double eval(Environment env) throws EvalException {
+		return env.get(pos, id);
+	}
 
 }
