+++ src/NodeStmtAssn.java
@@ -3,14 +3,14 @@
 
 public class NodeStmtAssn extends NodeStmt {
 
-    private NodeAssn assn;
+	private NodeAssn assn;
 
-    public NodeStmtAssn(NodeAssn assn) {
-	this.assn=assn;
-    }
+	public NodeStmtAssn(NodeAssn assn) {
+		this.assn = assn;
+	}
 
-    public int eval(Environment env) throws EvalException {
-	return assn.eval(env);
-    }
+	public double eval(Environment env) throws EvalException {
+		return assn.eval(env);
+	}
 
 }
